Dev-branch machines with Super I/O chips that will not be implemented in v2.10
Namely, the AMI Apollo and the Iwill P55XB2. Also, while it's Super I/O chip is implemented, the Packard Bell Bora Pro is being dev-branched as well for failing to POST.

Also adds a 66 MHz FSB Pentium II table in preparations for 440LX machines to be added in the future.