Add a Victor V86P disk driver
This emulates a JVC-branded controller/drive pair, using RLL encoding,
connected via a small connector electrically compatible with ST-506.

The controller is ST-506 compatible with an extra command for
self-power-off. The option ROM is made by SMS. Commented disassembly is
available for study [1].

The disk is a 3.5" 20MiB "made by Victor", labeled JD3824T100 on the outer
protective casing, JD3824T00-1 on the actual drive. It's 615/2/34
physically, pretends to be a 614/4/17 so that it's type 3 compatible.

[1] https://archive.org/details/v86p-hd