Update ASUS ISA-386C: Supports 64MB of RAM
According to documentation obtained here:
https://www.win3x.org/uh19/public/motherboard/manual/30218.pdf

The ASUS ISA-386C in addition to supporting 32MB onboard, supports an additional 32MB through a proprietary RAM card using an extended ISA connector marked "S1" on the board.

I have no idea on implementation details of this, but the manual seems to imply it just ends up mapped directly into RAM banks 2 and 3. So I'm not sure if raising the RAM limit here is the correct fix, but... yeah. Up to y'all. I won't be offended if you reject this PR out of hand :)