Device: Add device_get_instance
It is practical to store the instance number into the device context.
This way, during add_device_common->init call, the device init function
can ask its own instance number.

This can be used in the serial passthrough device, so that the serial port
number can be set equal to the instance number.

Signed-off-by: Andreas J. Reichel <webmaster@6th-dimension.com>