cmake: Fix Qt 6 plugin installation on macOS
Run find_package to import the CMake rules for the various plugins,
as install_qt5_plugin didn't work for me on Qt 6 otherwise, and install
the macOS style and icon plugins as well so that the widget style and
icons work correctly on Qt 6.

${_qt_plugin_dir} doesn't seem to get set by anything and
install_qt5_plugin hardcodes the PlugIns path, so hardcode that in
qt.conf as well.

Additionally, adjust the Qt library folder to take into account the
major version being used.