Revert "Fix crash at exit due to a unreleased mutex."
This reverts commit 80e547000673b1f8f9804a3cacbe5dc934077493.

std::unique_lock is incapable of recursively locking a mutex, which is needed for multi-monitor setups.
As a result it will crash/show undefined behaviour when switching renderers. Switch to instead calling
endblit() after pc_close to avoid crashes; at this point the CPU thread is now terminated so the mutex
no longer remains held by it.