Modem: Implement `ATD;` (return to command mode after dialing)
To simulate the in-progress dialing, the number before the semicolon
is appended to a temporary buffer; when an ATD command without a semicolon
is issued, the buffer contents are prepended and the complete number
is dialed at once.

Fixes Windows 98 dialer being stuck if "wait for dial tone" option
was enabled.