Add Micronics M5Pi
A 430LX PCI PhoneixBIOS Ax86 machine.with a relatively stock BIOS, a rarity. Also adds a single channel version of the W83769F to support it's specific IDE setup.. Do not use all 3 PCI slots and secondary IDE at the same time, that will cause an IRQ conflict as per the real board.