More ESP SCSI fixes (LUN mainly)
This commit should fix the NT 3.1 AMD PCscsi drivers when they look for devices on LUN > 0 and causing havoc with them (BSOD).
Basically clear the FIFO and set the SEQ_0 bit in the RSEQ read reg.
AM53c974 side: a biosless card is biosless.
General: fixed all the log excesses.