mach64: Split block 0 and block 1 register access routines
This is done to make supporting Rage XL easier, but I think it also
makes things a bit more straightforward to follow and does away with
some unnecessary & redundant bit masking.

On Rage XL, the registers are also mapped into a separate 4KB PCI memory
range (BAR3 -- the "auxiliary aperture"), optionally unmapping them from
the linear aperture, leaving the whole 8MB for the frame buffer memory.
However, the order of the two register blocks in the auxiliary aperture
is reversed: block 0 is on lower addresses (starting from zero) and
block 1 follows (starting from 0x400).

Let's separate the read/write callbacks for individual register blocks,
so that they could be conveniently wrapped in callbacks for the auxiliary
range, in a follow-up patch.