mach64: Add support for Rage XL's auxiliary range
On Rage XL, the registers are mapped into a separate 4KB PCI memory range
(BAR3 -- the "auxiliary aperture"), and optionally unmapped from the
linear aperture, leaving the full 8MB for the framebuffer.

In auxiliary aperture, the order of the two register blocks is reversed
compared to the linear apertue -- block 0 is on lower addresses
(starting from zero) and block 1 follows (starting from 0x400).