mach64: Add basic support for Rage XL
This chip is typically present on cheap cards and server motherboards.
It is a newer chip than the already supported VT2, but appears mostly
the same.

One significant difference is an addition of the auxiliary aperture (BAR3),
so that the whole linear aperture can be used for accessing the full 8MB
frame buffer. Support for this is added in a separate commit, to keep
clutter low.

The ROM image that has been tested to work well is this one:

  SHA1(sha1sum roms/video/mach64/XL_5219.bin)
  f87b2da3954fd4b55decbfb4a082e61d3e3e598b  roms/video/mach64/XL_5219.bin