Fix remember size and position regression in which vertical size could grow when reopening the VM (#6239)
* Change window resizing logic when using 4:3 aspect ratio

Change window resizing logic when using 4:3 aspect ratio to resize content only. Fixed issues in Remember size and position that are derived from the change. There is a slight flicker while resizing with force 4:3 aspect ratio.

* Fix regression that adds vertical size when showing toolbar and statusbar and remembering size and po

The recent rework on the force aspect 4:3 which needed to revisit the remembering of size and position introduced a regression that would save the size without considering the non-content window elements that could be showed or hidden. Now it calculates it accordingly.