Use lock emoji to indicate write-protected recent disk images.
This makes the menu look a little more visually distinct and obvious
when a recent disk image item is set to write-protected mode.  While
an icon does exist, I have found it to be small and visually
indistinguisable enough that I didn't even know it existed before I
dug into the code.

This prefixes image names with U+1F512 LOCK (🔒).  My first iteration
also added U+1F513 OPEN LOCK (🔓) to read-write image names, but I
quickly found that to be too visually distracting and confusing next
to the write-protect symbols, so I have opted to only keep the first
one.

This might have a caveat that the symbol will fail to show up on
systems without the lock emoji somewhere in their font set, but these
days, I expect pretty much everyone to be on OSes that come with the
full set of emoji fonts anyhow.